perf: Memo menu components
