+++ src/widgets/Menu/components/MenuLink.tsx
@@ -10,4 +10,4 @@ const MenuLink: React.FC<AnchorHTMLAttributes<HTMLAnchorElement>> = ({ href, ...
   return <Tag {...props} {...otherProps} />;
 };
 
-export default MenuLink;
+export default React.memo(MenuLink, () => true);
