+++ src/widgets/Menu/components/UserBlock.tsx
@@ -38,4 +38,4 @@ const UserBlock: React.FC<Props> = ({ account, login, logout }) => {
   );
 };
 
-export default UserBlock;
+export default React.memo(UserBlock, (prevProps, nextProps) => prevProps.account === nextProps.account);
